Classic mode almost finished

+++ src/main/resources/webroot/assets/js/boardList.js
@@ -1,6 +1,8 @@
 "use strict";
 document.addEventListener("DOMContentLoaded", init);
 let squareList;
+let turn;
+let turnOk;
 let color;
 let lItems;
 let currentSquare;
@@ -9,12 +11,14 @@ let pieceHolder;
 
 function init() {
     setupPage();
+    turn = "blue";
     lItems = document.getElementById("squareList").getElementsByTagName("li");
     lines = document.getElementById("squareList").getElementsByTagName("li");
     let squareList = document.getElementById('squareList');
     pieceHolder = document.getElementById('pieceHolder');
     localStorage.setItem("turn", "blue");
     document.querySelector("body h1").innerHTML = ("Blue goes first, don't look red!");
+    turnOk = true;
     setupClick();
 }
 
@@ -137,15 +141,21 @@ function setupPage() {
 }
 
 function setupClick() {
-    let squareList = document.getElementById('squareList');
-    squareList.onclick = function (e) {
-        deleteAllDots();
-        if (colorOfClick(e.target.id) === "blue") {
-            posmoves(e.target.id);
-        } else {
-            setupClick();
+    if (turn === "blue") {
+        let squareList = document.getElementById('squareList');
+        squareList.onclick = function (e) {
+            deleteAllDots();
+            if (turnOk) {
+                if (colorOfClick(e.target.id) === "blue") {
+                    posmoves(e.target.id);
+                } else {
+                    setupClick();
+                }
+            }
         }
-    };
+    } else {
+        getConfirm()
+    }
 }
 
 function posmoves(pieceName) {
@@ -155,8 +165,6 @@ function posmoves(pieceName) {
     let square = pieceName.split("-")[1]; // the position on the board
     square = parseInt(square, 10);
     color = colorOfClick(pieceName);
-
-    setupClick();
 
     if (name === "Bomb" || name === "Flag" || name === "lakeSquare" || name === "blankSquare")
         return; // if it's a piece that can't move
@@ -283,6 +291,8 @@ function activateDot(movedFromSquare, movedToSquare, type) {
         console.log(JSON.stringify(data));
         console.log(endCoordinates);
         getBoard();
+        sendTurn();
+        getConfirm();
     };
 }
 
@@ -383,7 +393,6 @@ function dotClicked(movedFromSquare, movedToSquare) {
         }
         flipPieces("blue");
         console.log(Switch);
-        getBoard();
         if (result === 1) {
             flipSinglePiece(newSquareID1);
         }
@@ -436,7 +445,6 @@ function recursive(movedToSquare, direction, movedFromSquare) {
     }
 }
 
-
 function combat(a, b) {
 // a is the attacking piece, if a wins the function returns 1, if b wins it returns -1, otherwise returns 0 if they tie both die, or 2 if its a flag
     if (a === b) // If they tie
@@ -487,20 +495,6 @@ function flipPieces(color) {
     }
 }
 
-function turnRed() {
-    for (let i = 0; i < 100; i++) {
-        let line = lines[i].innerHTML;
-        if (line.indexOf("red") !== -1) {
-            if (line.indexOf("red" + "Back") !== -1) {
-                let lineID = line.split("-")[0].split("id=\"")[1];
-                lines[i].innerHTML = line.replace(new RegExp("/(.*)png", "g"), pieces + lineID + ".png");
-            } else {
-                lines[i].innerHTML = line.replace(new RegExp("/(.*)png", "g"), pieces + "red" + "Back.png");
-            }
-        }
-    }
-}
-
 // a function to switch the backs of the pieces
 function flipSinglePiece(pieceName) {
 
@@ -540,7 +534,159 @@ function deleteAllDots() {
 }
 
 function getBoard() {
-    let data = {data: "nextTurn"};
+    fetch('/api/board').then(res => res.json()).then(function (response) {
+        let gameboard = Object.values(response);
+        let boardArray = [];
+        for (let i = 0; i < 10; i++) {
+            for (let j = 0; j < 10; j++) {
+                if (gameboard[0][i][j] === null) {
+                    boardArray.push(null);
+                } else {
+                    let code = gameboard[0][i][j].rank + "-" + gameboard[0][i][j].player;
+                    boardArray.push(code);
+                }
+            }
+        }
+        console.log(boardArray);
+        let color = "";
+        let boardLines = document.getElementById("squareList").getElementsByTagName("li");
+        for (let i = 0; i < 100; i++) {
+            let code = boardArray[i];
+            switch (code) {
+                case '3-1':
+                    code = '8';
+                    color = "red";
+                    break;
+                case '1-1':
+                    code = 'Spy';
+                    color = "red";
+                    break;
+                case '2-1':
+                    code = '9';
+                    color = "red";
+                    break;
+                case '6-1':
+                    code = '5';
+                    color = "red";
+                    break;
+                case '5-1':
+                    code = '6';
+                    color = "red";
+                    break;
+                case '7-1':
+                    code = '4';
+                    color = "red";
+                    break;
+                case '10-1':
+                    code = '1';
+                    color = "red";
+                    break;
+                case '4-1':
+                    code = '7';
+                    color = "red";
+                    break;
+                case '8-1':
+                    code = '3';
+                    color = "red";
+                    break;
+                case '9-1':
+                    code = '2';
+                    color = "red";
+                    break;
+                case '11-1':
+                    code = 'Bomb';
+                    color = "red";
+                    break;
+                case '0-1':
+                    code = 'Flag';
+                    color = "red";
+                    break;
+                case '3-2':
+                    code = '8';
+                    color = "blue";
+                    break;
+                case '1-2':
+                    code = 'Spy';
+                    color = "blue";
+                    break;
+                case '2-2':
+                    code = '9';
+                    color = "blue";
+                    break;
+                case '6-2':
+                    code = '5';
+                    color = "blue";
+                    break;
+                case '5-2':
+                    code = '6';
+                    color = "blue";
+                    break;
+                case '7-2':
+                    code = '4';
+                    color = "blue";
+                    break;
+                case '10-2':
+                    code = '1';
+                    color = "blue";
+                    break;
+                case '4-2':
+                    code = '7';
+                    color = "blue";
+                    break;
+                case '8-2':
+                    code = '3';
+                    color = "blue";
+                    break;
+                case '9-2':
+                    code = '2';
+                    color = "blue";
+                    break;
+                case '11-2':
+                    code = 'Bomb';
+                    color = "blue";
+                    break;
+                case '0-2':
+                    code = 'Flag';
+                    color = "blue";
+                    break;
+            }
+            if (boardArray[i] === null && i === 42 || 43 || 46 || 47 || 52 || 53 || 56 || 57) {
+                boardLines[i].innerHTML = "<div id=\"lakeSquare-" + i + "\"></div>";
+            }
+            if (boardArray[i] === null) {
+                boardLines[i].innerHTML = "<div id=\"blankSquare-" + i + "\"></div>";
+            } else {
+                boardLines[i].innerHTML = "<img src=\"../assets/media/pieces/" + color + code + ".png\" id=\"" +
+                    color + code + "-" + (i) + "\">";
+            }
+        }
+    })
+}
+
+function getConfirm() {
+    console.log('Retrieving messages...');
+    fetch('/api/nextTurn1').then(res => res.json()).then(function (response) {
+        console.log(response);
+        if (response === "Turn") {
+            console.log("Retrieving board.....");
+            getBoard();
+            turn = "blue";
+            turnOk = true;
+            setupClick();
+            fetch("/api/set2ToNull", {
+                method: "POST",
+                body: JSON.stringify({data: "null"}),
+                headers: {
+                    "Content-Type": "application/json"
+                }
+            })
+        }
+        setTimeout(getConfirm, 2000);
+    })
+}
+
+function sendTurn() {
+    let data = {data: "Turn"};
     fetch("/api/nextTurn2", {
         method: "POST",
         body: JSON.stringify(data),
@@ -550,9 +696,7 @@ function getBoard() {
     })
         .then(res => res.json())
         .then(json => console.log(JSON.stringify(json)));
-    console.log(JSON.stringify(data));
-    fetch('/api/board').then(res => res.json()).then(function (response) {
-        console.log(response);
-    });
+    turn = "red";
+    turnOk = false;
 }
